

\name{ConQuest.Log.Example1.log.bz2}

\docType{data}

\alias{ConQuest.Log.Example1.log.bz2}

\title{Example Log File from ConQuest}

\description{
This is a text file with the log from a ConQuest analysis
It can be accessed via bzfile ( file.path( .Library , "eat/extdata/ConQuest.Log.Example1.log.bz2" ) )
}

%\usage{data(ConQuest.Log.Example1.log.bz2)}

\format{txt}

%\source{}

%\references{
%}

%\keyword{}






