

\name{science1.context.vars}

\docType{data}

\alias{science1.context.vars}

\title{Science achievement test data - Context variable names}

\description{
This vector contains the names of context variables in data set \code{\link{science1}}
}

%\usage{data(science1.context.vars)}

\format{chr [1:9]}

%\source{}

%\references{
%}

%\keyword{datasets}






